use module level variable instead of settings directly
